Add Buffer, BufferCommand and printCommand, introduce views
Add  basic types  for  file buffer  and commands  operating  on it,  and
implement a  print command  using these.  To avoid  mixing IO  with pure
buffer editing, buffer commands return a Writer containing what would be
printed during interactive use.

To  avoid  repetitive  bound-checking,   buffer  commands  take  address
arguments guaranteed to be valid. This is enforced by making the Address
smart  constructors do  the  checks. Because  of that  we  want to  keep
constructors private, and so we introduce  View type class to be able to
pattern match on  addresses. ViewPatterns extension is  not used because
GHC  complains about  inexhaustive pattern  matches, and  we do  want to
check for that.