Fix leaky line address verification system
Editor commands  are functions  taking a  line address  and a  buffer on
which they will  run as arguments. To prevent from  running a command on
an invalid  address (for example  on a  line with negative  number), the
commands take  in values of  verified address  type which can  solely be
created with `verifyAddress` function.

However, it is  possible to construct an address against  one buffer and
use it  against another  one. This  commit fixes  that issue,  by making
address type hold the buffer against which it was verified.